+++ README
@@ -1,0 +1,7 @@
+
+
+1. Home page url:
+   /,/home,/index
+2. Top path for test page:
+   /test/
+   http protocol test page: /test/
