+++ marble/aiqiyi/spec/gen.sh
