+++ model/ops_interface.go
