--- model/xiaodu/gen.sh
@@ -1,57 +1,0 @@
-#!/bin/sh
-Machine=`uname -i`
-OutputDir=`pwd`"/.."
-UnionAiqiyi="aiqiyi.proto"
-RequestAiqiyi="bid_request.proto"
-ResponseAiqiyi="bid_response.proto"
-pkgName="aiqiyi.pb.go"
-function fixPackage() {
-    regxL1="const _ = proto.ProtoPackageIsVersion2 \/\/ please upgrade the proto package"
-    regxL2="func init() { proto.RegisterFile(\"aiqiyi.proto\", fileDescriptor0) }"
-    if [[ !(-f $1) ]];then
-        echo "$1 not exist, exiting..."
-        exit
-    fi
-    if [[  $Machine =~ "Mac" ]];then
-        echo $Machine" "$1
-        #pwd
-        sed -i "" "s/$regxL1/\/\/ $regxL1/g" $1
-        sed -i "" "s/$regxL2/\/\/ $regxL2/g" $1
-    else 
-        sed -i  "s/$regxL1/\/\/ $regxL1/g" $1
-        sed -i  "s/$regxL2/\/\/ $regxL2/g" $1
-        echo "In function $1"
-    fi
-}
-
-
-# Insert the following 2 lines to define
-# protobuf version and package information
-#     syntax = "proto2";
-#     package aiqiyi; 
-# and delete lines
-#     with extensions 100 to max;
-if [ -f $RequestAiqiyi ]; then
-    cat $RequestAiqiyi | sed   '/package ads_serving.proto;/i\
-    syntax = \"proto2\";\
-    package aiqiyi;\
-    ' | sed "5d" | sed /"extensions 100 to max;"/d> $OutputDir"/"$UnionAiqiyi
-else
-	echo $RequestAiqiyi not found!
-	exit
-fi
-if  [ -f $OutputDir"/"$UnionAiqiyi ]; then
-    echo "Generating $UnionAiqiyi request part successufflly!"
-    cat $ResponseAiqiyi | sed '1,4d'  | sed /"extensions 100 to max;"/d >>  $OutputDir"/"$UnionAiqiyi
-    cd $OutputDir
-    echo "Start generate package with .proto file $UnionAiqiyi."
-    protoc  --plugin=$GOPATH/bin/protoc-gen-go --go_out=$OutputDir $UnionAiqiyi
-if [[ -f $pkgName ]]; then
-    echo "Start fix generated go package file aiqiyi.pb.go."
-echo $(fixPackage $pkgName)
-fi
-else
-    echo "Generating $UnionAiqiyi request part failed! Exiting..." 
-    exit
-fi
-
