+++ template/home.htm
@@ -2,7 +2,7 @@
 <html>
     <head>
         <meta charset="utf-8">
-        <title> Toger Home</title>
+        <title>Garden from Mars</title>
         <link rel=stylesheet href="site.css">
     </head>
     <script type="text/javascript">
@@ -21,7 +21,7 @@
   </script>
 <body>
 <div class="example" id="writing-files">
-  <h2><a href="./t">Go template example</a></h2>
+  <h2><a href="./home">Golang Example Home</a></h2>
 <table>
 
 <tr>
@@ -270,12 +270,15 @@ we&rsquo;ve just seen to the <code>stdin</code> and <code>stdout</code> streams.
 </table>
   
   
-  <p class="next">
-  Next example: <a href="line-filters">Line Filters</a>.
+<p class="next">
+    List index: <a href="/list">list pages</a>.
 </p>
   
-  <p class="footer">
-  by <a href="https://twitter.com/mmcgrana">@mmcgrana</a> | <a href="mailto:mmcgrana@gmail.com">feedback</a> | <a href="https://github.com/mmcgrana/gobyexample/blob/master/examples/writing-files">source</a> | <a href="https://github.com/mmcgrana/gobyexample#license">license</a>
+<p class="footer">
+  by <a href="https://github.com/MarsBighead">Github MarsBighead</a> | 
+<a href="mailto:duanmhy@gmail.com">feedback</a> | 
+<a href="https://github.com/MarsBighead/garden">Garden Source</a> | 
+<a href="https://github.com/MarsBighead/garden/blob/master/LICENSE">license MIT</a>
 </p>
 </div>
   </body>
