+++ template/list.htm
@@ -1,0 +1,59 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="utf-8">
+        <title>Garden from Mars</title>
+        <link rel=stylesheet href="site.css">
+    </head>
+    <script type="text/javascript">
+    if (window.location.host == "gobyexample.com") {
+        var _gaq = _gaq || [];
+        _gaq.push(['_setAccount', 'UA-34996217-1']);
+        _gaq.push(['_trackPageview']);
+        (function() {
+            var ga = document.createElement('script'); 
+            ga.type = 'text/javascript'; ga.async = true;
+            ga.src = ('https:' == document.location.protocol ? 'https://ssl' : 'http://www') + '.google-analytics.com/ga.js';
+            var s = document.getElementsByTagName('script')[0]; 
+            s.parentNode.insertBefore(ga, s);
+        })();
+    }
+  </script>
+<body>
+<div class="example" id="writing-files">
+  <h2><a href="./list">List Example Home</a></h2>
+<table>
+    <tr>
+        <td class="docs">
+        <p>Writing files in Go follows similar patterns to the
+        ones we saw earlier for reading.</p>
+        </td>
+        <td class="code empty leading">
+        <a href="/json">JSON API Protocol</a>
+        </td>
+    <tr>
+        <td class="docs">
+        Example JSON Page 
+        </td>
+        <td class="code leading">
+        <a href="/json">JSON API Protocol</a>
+        </td>
+    </tr>
+</table>
+  
+  
+<p class="next">
+    List index: <a href="/list">list pages</a>.
+</p>
+  
+<p class="footer">
+  by <a href="https://github.com/MarsBighead">Github MarsBighead</a> | 
+<a href="mailto:duanmhy@gmail.com">feedback</a> | 
+<a href="https://github.com/MarsBighead/garden">Garden Source</a> | 
+<a href="https://github.com/MarsBighead/garden/blob/master/LICENSE">license MIT</a>
+</p>
+</div>
+  </body>
+
+</html>
+
