+++ template/statistics.htm
@@ -6,7 +6,7 @@
     </head>
     <body><h3>Statistics</h3>
     <p>Computes basic statistics for a given list of numbers</p>
-    <form action="localhoyts:8080/" method="POST">
+    <form action="localhoyts:8001/" method="POST">
         <label for="numbers">Numbers (comma or space-separated):</label><br />
         <input type="text" name="numbers" size="30"><br />
         <input type="submit" value="Calculate">
